Refactor: GptService 분리 및 config 삭제
